backend/requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.7.1
pymongo==4.16.0
bcrypt==5.0.0
PyJWT==2.11.0
python-dotenv==1.2.1
twilio==9.10.2
groq==0.4.1
requests==2.32.5
gunicorn==22.0.0